requirements.txt
tensorflow==2.17.0  
Pillow==10.4.0
opencv-python==4.10.0.84
Flask==3.0.3
numpy==1.24.3
scipy==1.11.1
matplotlib==3.9.4
seaborn==0.13.2
scikit-learn==1.3.0 
Werkzeug==2.3.7
gunicorn==21.2.0

